==> 01_VideoCollector.ipynb <==
import requests
import json
import time
import numpy as np
import pandas as pd
import os

import scanning_lib 

# Столбцы таблиц с видео
column_names_video = ['id', 
                      'title',  
                      'publishedAt', 
                      'channelTitle', 
                      'viewCount', 
                      'likeCount', 
                      'dislikeCount', 
                      'commentCount']

# Список каналов на скан:
channelId_list = ["UCUGfDbfRIx51kJGGHIFo8Rw", # Кац
                  "UCgxTPTFbIbCWfTR9I2-5SeQ", # Навальный лайф
                  "UCdubelOloxR3wzwJG9x8YqQ", # Дождь
                  "UCxT6lMsQq4uC7yVfN6joLYA" # Наливкин
                 ] 
# Ограничение по дате выпуска 
time_lower_bound = "2019-01-01T01:01:01Z"

# Папка куда скидывать таблицы с видео
storage_dir_name = "RawChannelVideos"

# Ключи Youtube Data APIv3
api_keys = ["AIzaSyB1S2CxQUAJp6oPbfhXXti7gmtGlR9rn5k", 
            "AIzaSyBRxQBh9kxW6MeyQViG3hvPqpJ-Jeza7_g", 
            "AIzaSyD54Pl9rqRmXTmV0UIK-I2trL0dKtSakqo", 
            "AIzaSyA20UizUidh_LltLuo3_rZJIBHl9XXa7vA", 
            "AIzaSyBBLMXx6kc2Djlc5L_YAtrAfUhnyUMmMsw", 
            "AIzaSyCPscCvM_mHmn0EGjW5wOj-sJcUf5x0JOE", 
            "AIzaSyCj5wdYLiKzzJfdyDyv15yp1jakzpJP-I8", 
            "AIzaSyCONGOiT5SPTI5mJk8YNRV7ekkWQ7A8ESg", 
            "AIzaSyB4vdPD_GVFuH3-Sf6Qy3MlzxqNvThkTjg"]

api_key = "AIzaSyCPscCvM_mHmn0EGjW5wOj-sJcUf5x0JOE"

# Вся процедура для 1 видео
def save_video_list(channelId, time_lower_bound, storage_dir_name, api_key):
    video_records = scanning_lib.scan_single_channel(channelId, time_lower_bound, api_key)


    if not os.path.isdir(storage_dir_name):
        os.mkdir(storage_dir_name)
    df_videos = pd.DataFrame(data=video_records, columns=column_names_video)
    video_file_name = storage_dir_name + """/""" + channelId + "_channel_videos.csv"
    df_videos.to_csv(video_file_name)
    return 0

# Собираем таблицы с видео
for api_key in api_keys:
    try:
        print("Start using new key")
        for channelId in channelId_list:
            ans = save_video_list(channelId, time_lower_bound, storage_dir_name, api_key)
            print(ans)
        break
    except:
        print("Change key")
        continue
print("Video collection completed")

# Отфильтруем видосы по диапазону дат более детально
video_files = [storage_dir_name + """/""" + file_name for file_name in os.listdir(storage_dir_name) 
                 if file_name.find("_channel_videos.csv") >= 0]
video_files

# Более тонкая фильтрация по датам
video_file = video_files[1]
df_container = []
for video_file in video_files:
    df_video= pd.read_csv(video_file)[column_names_video]
    df_container.append(df_video[(pd.to_datetime(df_video["publishedAt"], format="%Y-%m-%dT%H:%M:%SZ") >= '2020-05-01') \
                               & (pd.to_datetime(df_video["publishedAt"], format="%Y-%m-%dT%H:%M:%SZ") <= '2020-06-26')])
df_videos = pd.concat(df_container)
df_videos.head()

# Сохраняем
df_videos.to_csv('videos_to_analyze.csv')
