==> 02_CommentsCollector.ipynb <==
import requests
import json
import time
import numpy as np
import pandas as pd
import os
import re

# Самописные функции
import scanning_lib

# Столбцы таблиц с видео
column_names_video = ['id', 
                      'title',  
                      'publishedAt', 
                      'channelTitle', 
                      'viewCount', 
                      'likeCount', 
                      'dislikeCount', 
                      'commentCount']
# Столбцы таблиц с комментами
column_names_comment = ["id",
                        "videoId",
                        "authorVideoName",
                        "totalReplyCount",
                        "likeCount",
                        "authorDisplayName",
                        "textDisplay",
                        "publishedAt", "isReply"]

# Папка куда скидывать таблицы с комментами
storage_dir_name = "RawComments"

# Файд откуда подгружать список видео
videos_table = "videos_to_analyze.csv"

# Ключи Youtube Data APIv3 
api_keys = ["AIzaSyB1S2CxQUAJp6oPbfhXXti7gmtGlR9rn5k", 
            "AIzaSyBRxQBh9kxW6MeyQViG3hvPqpJ-Jeza7_g", 
            "AIzaSyD54Pl9rqRmXTmV0UIK-I2trL0dKtSakqo", 
            "AIzaSyA20UizUidh_LltLuo3_rZJIBHl9XXa7vA", 
            "AIzaSyBBLMXx6kc2Djlc5L_YAtrAfUhnyUMmMsw", 
            "AIzaSyCPscCvM_mHmn0EGjW5wOj-sJcUf5x0JOE", 
            "AIzaSyCj5wdYLiKzzJfdyDyv15yp1jakzpJP-I8", 
            "AIzaSyCONGOiT5SPTI5mJk8YNRV7ekkWQ7A8ESg", 
            "AIzaSyB4vdPD_GVFuH3-Sf6Qy3MlzxqNvThkTjg"]

# Собираем и сохраняем комменты с одного видео
def save_comment_list(video_Id, author_video_name, storage_dir_name, api_key):
    record_strings = scanning_lib.scan_single_video(video_Id, author_video_name, api_key)

    if not os.path.isdir(storage_dir_name):
        os.mkdir(storage_dir_name)
        
    df_comments = pd.DataFrame(data=record_strings, columns=column_names_comment)
    comment_file_name = storage_dir_name + """/""" + video_Id + "_video_comments.csv"
    
    df_comments.to_csv(comment_file_name)
    return 0

# Сохраняем комменты для каждого видео в отдельный файл
for api_key in api_keys:
    
    df_fvideos = pd.read_csv(videos_table)[column_names_video]
    
    comment_files = [storage_dir_name + """/""" + file_name for file_name in os.listdir(storage_dir_name) 
                     if file_name.find("_video_comments.csv") >= 0]
    existed_v_ids = [re.split(storage_dir_name + r'\/(.*)_video_comments.csv', comment_file)[1] 
                             for comment_file in comment_files]
    
    try: 
        print("Start using new key")
        for record in df_fvideos.T.iteritems():
            video_Id = record[1]["id"]
            ans = "Already scanned"
            if not any([existed_v_id == video_Id for existed_v_id in existed_v_ids]):
                author_video_name = record[1]["channelTitle"]
                ans = save_comment_list(video_Id, author_video_name, storage_dir_name, api_key)
            print(ans)
        break
    except:
        print("Change key")
        continue
print("Collection completed")
